feat(datadog-opentelemetry): add a span exporter that uses the libdatadog trace exporter (#8)
# What does this PR do?

* Add a `datadog-opentelemetry` crate in the workspace
* This crate contains a `DatadogSpanExporter` which sends otel spans to
the trace agent using the TraceExporter
* This crate also contains code doing (an incomplete see TODOs)
conversion from otel span data to datadog spans
  * An example of setting up the "datadog distribution" of the otel sdk

The integration of the TraceExporter in the SpanExporter is a bit
complex, and can probably be optimized.
We spawn a new thread, and send span data to be sent or shutdown
requests through a channel. The worker then sends back send results back
to the span exporter though another channel.
This is done because the current APIs exposed for
SpanProcessor/SpanExporter in otel sdks are pretty bad and lead to
crashes/deadlocks if used improperly within an async context.
(See comment on `DatadogExporter`)

# Motivation

What inspired you to submit this pull request?

# Additional Notes

Anything else we should know when reviewing?

[APMSP-1932]

[APMSP-1932]:
https://datadoghq.atlassian.net/browse/APMSP-1932?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ