feat(configuration): add configuration from sources
Improve configuration to read from environment variables, and provide an override API.

We can configure a list of sources at config creation.  We also collect all errors encoutered until the first correct value.
It also returns the origin of the configuration for reporting once we have configuration telemetry, although this is currently discarded.