feat(datadog-opentelemetry): add a Datadog SpanProcessor instead of a SpanExporter (#19)
# Motivations

To attach the sampling decision correctly to traces, we need to flush
the whole trace chunk together with the root chunk. The SpanExporter
gets passed SpanData as it arrives, without any groupping by Trace,
which mean we cannot use it.

# Changes

This PR modifies the datadog components in the following ways:
* Add a TraceRegistry: This struct bookeeps information while the tarce
is active, and buffers all closed spans until the full trace is finished
* Add a DatadogSpanProcessor: Registers the span in the registry, and
adds the trace chunk in a batch once it's done
* Modify the DatadogSpanExporter: This component no longer implements
the opentelemetry_sdk::SpanExporter trait, because the API it provides
is not suitable. Instead it now keeps trace chunks whole when exporting
them. It also does buffering and flushes based on batch size and age.

# Buffering behaviour

Finished spans are now kept in memory in one of 3 places:
* In the registry if the trace is still not completely finished
* In the Exporter batch buffer
* In the TraceExporter after we have grabbed a batch, but haven't
dropped it yet

[APMSP-1979]

[APMSP-1979]:
https://datadoghq.atlassian.net/browse/APMSP-1979?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ