feat(sampler): integrate sampler with the text map propgator
This PR refactors the sampler a bit in order to store thre sampling decision in the TraceRegistry.
The `ShouldSample` implementation is know in the datadog-opentelemetry crate, and the DatadogSampler uses simpler code.

To test that this works, this PR adds integration tests with the test agent that create a new span, and send it. The snapshots should contain the expected sampling tags, and we check that the tracecontext contains expected keys.