feat: unify the definitions for sampling decision/mechanism in dd-trace
We have the definitions for sampling decision and mechanism in three different place, the sampler, the extractor and the registry.
This PR replaces all three of them with structs in dd-trace. This struct is explicitely not an enum but a number so we can handle and propagate unknown values correctly