feat(sampling): unify the definitions for sampling decision and mechanism in dd-trace (#30)
# What does this PR do?

We have the definitions for sampling decision and mechanism in three
different place, the sampler, the extractor and the registry. This PR
replaces all three of them with structs in dd-trace. This struct is
explicitely not an enum but a number so we can handle and propagate
unknown values correctly

# Motivation

What inspired you to submit this pull request?

# Additional Notes

Anything else we should know when reviewing?