fix(propagation): add tracestate to datadog context (#27)
# What does this PR do?

_If the tracecontext propagator is configured, even if it is not the
first propagator to extract the trace context, the tracestate will be
saved in the local trace context if the traceparent trace-id matches the
extracted the trace-id._

At the moment, only the `tracestate` tag was set. This PR sets also
primary_context tracestate property in the datadog's extracted context.

# Additional Notes
It will fix some ST from
[test_headers_preference.py](https://github.com/DataDog/system-tests/blob/main/tests/parametric/test_headers_precedence.py#L693)