feat(otel_mapping): unify the otel to dd mapping code (#31)
# What does this PR do?

The OTEL to DD span mapping code was duplicated in both the
datadog-opentelemetry crate and the sampler crate.
This PR extracts this code to a new crate, which both crate import and
removes the code duplication in the sampler.

To make the code generic enough, this modifies the sampler function to
pass a `PreSampledSpan` which bundles attributes and otel resource in
one object implementing the same trait as the Span we pass on export.

This PR shouldn't modify the behaviour of the code.

# Motivation

What inspired you to submit this pull request?

# Additional Notes

Anything else we should know when reviewing?