fix(sampling): updates to make system-tests pass (#39)
# What does this PR do?

Service name not being set in a way that's configurable for the span.
This allows us to vary service name per span which is a necessity for
system-tests.

Fixes us setting improper tag '_dd.rule.psr' when we should be setting
’_dd.rule_psr’

# Motivation

Make Datadog parametric system-tests that don't rely on lazy sampling
pass.

# Additional Notes

Anything else we should know when reviewing?

---------

Co-authored-by: Igor Unanua <igor.unanua@datadoghq.com>