feat(sampler): propagate the agent response back to the sampler to sample by agent rates (#35)
# What does this PR do?

This PR exposes a callback from the trace exporter into the sampler to
return agent rates.

The agent rates is in a RwLock because it need to be periodically
sampled.

This PR also depends on an unmerged branch of the trace exporter which
needs to be updated once the linked PR is merged.
https://github.com/DataDog/libdatadog/pull/1104

# Motivation

What inspired you to submit this pull request?

# Additional Notes

Anything else we should know when reviewing?