feat(remote_config): make the remote config worker shutdown when the tracer stops (#67)
# What does this PR do?

* Split the RemoteConfigClient in two. Just the client to perform
related to fetching and applying remote config changes, and a
RemoteConfigClientWorker that spawns a thread to perform the fetch loop.
Whenever we create a worker, we get back a handle that can be used to
trigger and wait for shutdown.
The worker also waits on a cancellation tokens, and signals back
whenever it shuts down.

* Refactor a bit the client itself. While trying to understand the code,
I found a few ways to make it easier to read.

* Add integration tests for remote config using the trace agent

* Refactor some of the logic related to the SpanProcessor shutdown. Now
we first trigger both the exporter and rc worker shutdown, and then wait
for them in order

# Motivation

Whenever we create a new Tracer, we also create a RC background worker.
These workers are never stopped, and since they hold callbacks into the
sampler, they keep quite a lot of memory alive forever which could cause
memory leaks.
Also there are no integration tests using RC rules.

# Additional Notes

Anything else we should know when reviewing?