feat(config): remove mutex around config and instead use arc swap
# Changes

* Remove the mutex around the config we pass around and instead use ArcSwap to swap the remote config
* unify the config origin struct between config and source files

# Motivation

We are currently passing an Arc<Mutex<Config>> around, and because the config itself contains an Arc<Mutex<Callbacks>>
That seems like a recipe for a deadlock if you try to modify the config while inside the callback which you accessed by locking the config lock.

The Mutex is here so that the config can be updated for RC, but we can instead wrap it in an ArcSwap, which is lockless.
Since the configs are written to by a single entity, and we don't need a linear view of it that will be fine, and probably more performant on concurrent access