feat(span registry): fix span registration when using the otel API (#75)
# Motivations

This PR fixes four bugs:

* When using the tracing API, since Otel spans are only materialised
when they are closed, and not as soon as they are started we had issues
with the TraceRegistry. Single spans would be opened and closed, without
having a parent registered and then flushed since they are the only span
in the trace chunk we know about. Since, in this order, the root span is
registered last we miss all of the context propagation data and they
would not inherit the sampling decision.
* We did not tag the top level spans properly. The otel mapping code has
some logic for it, which is wrong. And due to how the TraceExporter
compute stats, we would drop traces that have a positive sampling
decision....
* If the decision maker was missing from the incoming context, we would
tag it with dm=-3 (local rule) which is wrong.
* If the sampling decision was missing from the incoming context, and
there was some Datadog style context we would default to user keep. This
is wrong and we should instead sample the span ourselves

# What does this PR do?

Here is how this is fixed:
* We register the local root spans in the sampler, instead of in
on_start. The tracing-otel bridge does call the sampler for every span
before creating a child, which means we can register spans properly
* We don't set top level ourselves and instead rely on the TraceExporter
to do it properly for us
* We carry an Option<Mechanism> everywhere we store sampling info for a
trace chunk, instead of having to pick an incorrect value

Furthermore this PR adds integration tests that call to the library
through tracing, testing context extraction and injection to make sure
there are no regressions