feat(benches): add scaffolding for memory allocation benches
# What does the PR do

This PR adds a criterion Measurement which can be used to report the number of allocations performed during a benchmarks.

Memory benches are fairly different from normal wall time benchmarks because they are (fairly) deterministic.
We run the benchmark for a lot less time and iterations, and have to add a small epsilon to the measures so criterion doesn't explode.
But otherwise this works well.

Benchmarks can be made generic on the measurement to run both for wall_time and allocated bytes.