feat(otel_mappings): reduce memory allocations
# Motivation

The otel mapping code allocates a lot.
It wraps everything in a ByteString, which causes non static data to go in an Arc, furthermore all string fields on the otel spans needed to be copied to have a static lifetime.
This PR tries to remediate to that by borrowing more data instead of copying a reallocating.

# What does this PR do

This PR does multiple things:
* Change the SpanText implementation from ByteString which needs to own the data to a Cow<'a, str> where 'a is the lifetime of the otel span from which the dd span is derived.
This means that we can now borrow most of the string fields from the otel span, with the exceptions of new string/field values and attributes we have to cast to string.
* Preallocate maps to be roughly the size they are going to be to prevent small sizes re-allocations when we grow them

The tests also had to be modified a bit since we now need to keep the original spans alive