feat(trace-registry): optimize access
# Motivations

The trace registry is currently a shared data structure protected by a single lock.
This means that 2 threads accessing different traces will create contention.

The goal of this PR is to optimize access to the trace registry so creating and finishing spans can be as fast as possible and not block the user critical path.

# Changes

The optimisations work in two ways:
* For single threaded access: The default std Hashmap has been replaced by hasbrown. hasbrown uses foldhash, a non cryptographic hash function which is ~10x faster than the std hash function
* For concurrent access: The trace registry now shards the lock based on the trace id.
This means that we divide the trace registry in 64 locks. So two different traces now only have 1/64 chances of colliding.

Through testing I have figured out that this should allow us to scale to up to 32 threads doing frequent concurrent reads on the trace registry