perf(injection): precompute the injection tag size (#94)
# What does this PR do?

This saves us reallocation of the String, as if we add only the _dd.p.dm
tag and _dd.p.tid tags, we are at least at more than 32 bytes. Since the
Vec smallest allocation is 8 bytes, and doubles on growth than means
were perform probably 2 re-allocations in the best case scenario.