feat(config_inversion): [APMAPI-1627] Centralized configuration registry basic implementation (#101)
# What does this PR do?

This implements the basis of centralized configuration registry (FKA as
config inversion) by adding a script to generate the
`supported_configurations.rs` file from the
`supported-configurations.json` file, and use the resulting enums
instead of strings for accessing env vars.

# Motivation

As dd-trace-rs is not used by customers yet, it is a great opportunity
to ensure we correctly document each configuration in it.

# Additional Notes

This does NOT implement yet: aliases, deprecations, handling of type,
default values and propertyKeys.

---------

Co-authored-by: Benjamin De Bernardi <debernardi.benjamin@gmail.com>