feat(span_processor): add debug mode for abandoned spans
# Motivation

This feature is inspired from the go tracer https://github.com/DataDog/dd-trace-go/pull/2188

It's hard to understand and locate spans that are never finished.
This PR adds a debug mode to the tracer that will track the age and root span name of traces, and
* log warnings from time to time if the traces are older than some amount of time.
* log warning if some traces are still open during tracer shutdown

# Implementation

This feature should not add any cost to the tracer if it is not enabled, but requires storing some extra data associated with each trace.

In order to not use any extra memory if the feature is not enabled, I track it in an additional registry which is only used in debug mode.

This registry tracks the root span name, and the age of the trace.

The debug mode is controlled by 2 new configurations:
* DD_TRACE_DEBUG_OPEN_SPANS
* DD_TRACE_OPEN_SPAN_TIMEOUT

In order to test this feature correctly, I added extra code to intercept and store logs during integration tests.
This is done through a thread local Logger, which is overridden and propagated locally during tests.

Everything is hidden behind the test-utils feature and should thus be zero cost