chore(ci): move system-tests branch back to main after upgrade (#110)
# What does this PR do?

Change the pinned hash of system-tests

# Motivation

In a previous PR (#109) I changed the system-test pin to point to a dev
branch, in order to validate non backward compatible changes. The dev
branch has been merged in main and has been deleted.
So the pinned commit is not valid anymore, and we should use the pin to
the latest commit of the main branch of system-tests.