chore(versioning): add release job (#114)
# What does this PR do?

Add a publish workflow which will be triggered by tag. The tag must have
the follwing format: `<crate_name>-vx.y.z`. The job is part of the
`publish` environment which means it will be held until a member of
`apm-rust` approves it. Upon approval, the job will perform some sanity
checks on the crate and then publish it to crates.io.

---------

Co-authored-by: Igor Unanua <igor.unanua@datadoghq.com>