feat(api): refactor public api visibility (#121)
This PR breaks backward compatibility since we remove previously exported items

# Changes
* Mark modules we use in benchmarks as pub only if we have the test-utils feature.
* Re-export log and configuration publicly modules from the top level to hide module structure
* Mark remote config update related function as pub(crate) since they should not be called by users
* Remove the deprecated "init_datadog" method to create the tracer

# Motivation

We should be very deliberate in what APIs we want to expose. Exposing more than necessary will lead customers to rely on implementation details. For this purpose, we should try to not expose the current module structure of datadog-opentelemetry and put items at their "natural" import paths.

To explore what function is visible, "cargo doc" is a great tool


Co-authored-by: paul.legranddescloizeaux <paul.legranddescloizeaux@datadoghq.com>